code cleanup #5
