+++ Common/ModSystems/DownedBossSystem.cs
@@ -1,5 +1,4 @@
-﻿using Consolaria.Common.ModSystems;
-using Consolaria.Content.Items.Consumables;
+﻿using Consolaria.Content.Items.Consumables;
 using Consolaria.Content.NPCs.Bosses.Lepus;
 using Microsoft.Xna.Framework;
 using System.Collections.Generic;
@@ -11,8 +10,9 @@ using Terraria.ID;
 using Terraria.ModLoader;
 using Terraria.ModLoader.IO;
 
-namespace Consolaria.Common.ModSystems {
-	public class DownedBossSystem : ModSystem {
+namespace Consolaria.Common.ModSystems
+{
+    public class DownedBossSystem : ModSystem {
 		public static bool downedLepus = false;
 		public static bool downedTurkor = false;
 		public static bool downedOcram = false;
