+++ Content/EmoteBubbles/McMoneyPantsEmote.cs
@@ -1,10 +1,9 @@
-using Microsoft.Xna.Framework;
 using Terraria.GameContent.UI;
 using Terraria.ModLoader;
 
 namespace Consolaria.Content.EmoteBubbles
 {
-	public class McMoneyPantsEmote : ModEmoteBubble
+    public class McMoneyPantsEmote : ModEmoteBubble
 	{
 		public override void SetStaticDefaults() {
 			AddToCategory(EmoteID.Category.Town);
