+++ Content/EmoteBubbles/TurkorEmote.cs
@@ -1,11 +1,10 @@
 using Consolaria.Common.ModSystems;
-using Microsoft.Xna.Framework;
 using Terraria.GameContent.UI;
 using Terraria.ModLoader;
 
 namespace Consolaria.Content.EmoteBubbles
 {
-	public class TurkorEmote : ModEmoteBubble
+    public class TurkorEmote : ModEmoteBubble
 	{
 		public override void SetStaticDefaults() 
 		{
