+++ Content/Items/Consumables/OcramBag.cs
@@ -1,7 +1,11 @@
+using Consolaria.Content.Items.Accessories;
+using Consolaria.Content.Items.Materials;
 using Consolaria.Content.Items.Weapons.Magic;
 using Consolaria.Content.Items.Weapons.Melee;
 using Consolaria.Content.Items.Weapons.Ranged;
 using Consolaria.Content.Items.Weapons.Summon;
+
+using Consolaria.Content.Items.Vanity;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using Terraria;
@@ -10,13 +14,12 @@ using Terraria.GameContent.ItemDropRules;
 using Terraria.ID;
 using Terraria.ModLoader;
 
-namespace Consolaria.Content.Items.BossDrops.Turkor {
-    public class TurkorBag : ModItem {
+namespace Consolaria.Content.Items.Consumables
+{
+    public class OcramBag : ModItem {
         public override void SetStaticDefaults () {
 
             ItemID.Sets.BossBag [Type] = true;
-            ItemID.Sets.PreHardmodeLikeBossBag [Type] = true;
-
             Item.ResearchUnlockCount = 3;
         }
 
@@ -27,7 +30,7 @@ namespace Consolaria.Content.Items.BossDrops.Turkor {
             Item.maxStack = 9999;
             Item.consumable = true;
 
-            Item.rare = ItemRarityID.Blue;
+            Item.rare = ItemRarityID.Lime;
             Item.expert = true;
         }
 
@@ -35,12 +38,16 @@ namespace Consolaria.Content.Items.BossDrops.Turkor {
             => true;
 
         public override void ModifyItemLoot (ItemLoot itemLoot) {
-            itemLoot.Add(ItemDropRule.OneFromOptionsNotScalingWithLuck(1, new int [] { ModContent.ItemType<FeatherStorm>(), ModContent.ItemType<GreatDrumstick>(), ModContent.ItemType<TurkeyStuff>() }));
-            itemLoot.Add(ItemDropRule.Common(ModContent.ItemType<SpicySauce>(), 2, 20, 39));
-            itemLoot.Add(ItemDropRule.Common(ModContent.ItemType<TurkorMask>(), 8));
+            itemLoot.Add(new OneFromRulesRule(1, ItemDropRule.Common(ModContent.ItemType<EternityStaff>()),
+                ItemDropRule.Common(ModContent.ItemType<DragonBreath>()),
+                ItemDropRule.Common(ModContent.ItemType<OcramsEye>()),
+                ItemDropRule.Common(ModContent.ItemType<Tizona>())));
 
-            itemLoot.Add(ItemDropRule.Common(ModContent.ItemType<HornoPlenty>()));
-            itemLoot.Add(ItemDropRule.CoinsBasedOnNPCValue(ModContent.NPCType<NPCs.Bosses.Turkor.TurkortheUngrateful>()));
+            itemLoot.Add(ItemDropRule.Common(ModContent.ItemType<OcramMask>(), 8));
+            itemLoot.Add(ItemDropRule.Common(ModContent.ItemType<SoulofBlight>(), 1, 25, 40));
+
+            itemLoot.Add(ItemDropRule.Common(ModContent.ItemType<ShadowboundExoskeleton>()));
+            itemLoot.Add(ItemDropRule.CoinsBasedOnNPCValue(ModContent.NPCType<NPCs.Bosses.Ocram.Ocram>()));
         }
 
         public override Color? GetAlpha (Color lightColor)
@@ -91,7 +98,7 @@ namespace Consolaria.Content.Items.BossDrops.Turkor {
 
             for (float i = 0f; i < 1f; i += 0.34f) {
                 float radians = (i + timer) * MathHelper.TwoPi;
-                spriteBatch.Draw(texture, drawPos + new Vector2(0f, 4f).RotatedBy(radians) * time, frame, new Color(165, 42, 42, 75), rotation, frameOrigin, scale, SpriteEffects.None, 0);
+                spriteBatch.Draw(texture, drawPos + new Vector2(0f, 4f).RotatedBy(radians) * time, frame, new Color(138, 43, 226, 80), rotation, frameOrigin, scale, SpriteEffects.None, 0);
             }
             return true;
         }
