+++ Content/Items/Pets/CursedFang.png
