+++ Content/Items/Placeable/EerieOcramMusicBox.cs
@@ -2,7 +2,7 @@ using Terraria.ModLoader;
 using Terraria;
 using Terraria.ID;
 
-namespace Consolaria.Content.Items.BossDrops.Ocram {
+namespace Consolaria.Content.Items.Placeable {
     public class EerieOcramMusicBox : ModItem {
 		public override void SetStaticDefaults () {
 			Item.ResearchUnlockCount = 1;
