+++ Content/Items/Placeable/OcramRelic.png
