+++ Content/Items/Placeable/OcramTrophy.cs
@@ -4,7 +4,7 @@ using Terraria.GameContent.Creative;
 using Microsoft.Xna.Framework;
 using Terraria.ID;
 
-namespace Consolaria.Content.Items.BossDrops.Ocram
+namespace Consolaria.Content.Items.Placeable
 {
     public class OcramTrophy : ModItem
     {
