+++ Content/Items/Placeable/TurkorRelic.png
