+++ Content/Items/Vanity/TurkorMask.cs
@@ -1,13 +1,13 @@
 using Terraria.ModLoader;
 using Terraria;
 using Terraria.ID;
-using Terraria.GameContent.Creative;
 using Microsoft.Xna.Framework;
+using Terraria.GameContent.Creative;
 
-namespace Consolaria.Content.Items.BossDrops.Ocram
+namespace Consolaria.Content.Items.Vanity
 {
 	[AutoloadEquip(EquipType.Head)]
-	public class OcramMask : ModItem
+	public class TurkorMask : ModItem
 	{
 		public override void SetStaticDefaults() {
 			CreativeItemSacrificesCatalog.Instance.SacrificeCountNeededByItemId[Type] = 1;
