+++ Content/NPCs/Bosses/Ocram/Ocram.cs
@@ -1,11 +1,14 @@
 using Consolaria.Common;
 using Consolaria.Common.ModSystems;
-using Consolaria.Content.Items.BossDrops.Ocram;
+using Consolaria.Content.Items.Consumables;
 using Consolaria.Content.Items.Materials;
 using Consolaria.Content.Items.Weapons.Magic;
 using Consolaria.Content.Items.Weapons.Melee;
 using Consolaria.Content.Items.Weapons.Ranged;
 using Consolaria.Content.Items.Weapons.Summon;
+using Consolaria.Content.Items.Pets;
+using Consolaria.Content.Items.Placeable;
+using Consolaria.Content.Items.Vanity;
 using Consolaria.Content.Projectiles.Enemies;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
@@ -14,14 +17,14 @@ using System.Collections.Generic;
 using Terraria;
 using Terraria.Audio;
 using Terraria.Chat;
-using Terraria.DataStructures;
 using Terraria.GameContent.Bestiary;
 using Terraria.GameContent.ItemDropRules;
 using Terraria.ID;
 using Terraria.Localization;
 using Terraria.ModLoader;
 
-namespace Consolaria.Content.NPCs.Bosses.Ocram {
+namespace Consolaria.Content.NPCs.Bosses.Ocram
+{
     [AutoloadBossHead]
     public class Ocram : ModNPC {
         public static LocalizedText BestiaryText {
