+++ Content/Projectiles/Friendly/Pets/OldLady.cs
@@ -1,13 +1,13 @@
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using System;
-using System.Linq;
 using Terraria;
 using Terraria.Audio;
 using Terraria.ID;
 using Terraria.ModLoader;
 
-namespace Consolaria.Content.Projectiles.Friendly.Pets {
+namespace Consolaria.Content.Projectiles.Friendly.Pets
+{
     public class OldLadyData : ModPlayer {
         public bool IsOldLadyPetActive { get; internal set; }
