+++ Content/Projectiles/Friendly/Pets/Worm.cs
@@ -1,5 +1,4 @@
 using Microsoft.Xna.Framework;
-using Microsoft.Xna.Framework.Graphics;
 using System;
 using System.Linq;
 using Terraria;
@@ -7,7 +6,8 @@ using Terraria.Audio;
 using Terraria.ID;
 using Terraria.ModLoader;
 
-namespace Consolaria.Content.Projectiles.Friendly.Pets {
+namespace Consolaria.Content.Projectiles.Friendly.Pets
+{
     public class WormData : ModPlayer {
         public bool IsWormPetActive { get; internal set; }
