+++ Content/Tiles/LepusRelic.cs
@@ -1,10 +1,6 @@
-using Terraria;
-using Terraria.DataStructures;
-using Terraria.ModLoader;
-
 namespace Consolaria.Content.Tiles
 {
-	public class LepusRelic : BossRelic
+    public class LepusRelic : BossRelic
 	{
 		public override string RelicTextureName => "Consolaria/Content/Tiles/LepusRelic";
 	}
