+++ Content/Tiles/OcramRelic.cs
@@ -1,10 +1,6 @@
-using Terraria;
-using Terraria.DataStructures;
-using Terraria.ModLoader;
-
 namespace Consolaria.Content.Tiles
 {
-	public class OcramRelic : BossRelic
+    public class OcramRelic : BossRelic
 	{
 		public override string RelicTextureName => "Consolaria/Content/Tiles/OcramRelic";
 	}
