+++ Content/Tiles/TurkorRelic.cs
@@ -1,10 +1,6 @@
-using Terraria;
-using Terraria.DataStructures;
-using Terraria.ModLoader;
-
 namespace Consolaria.Content.Tiles
 {
-	public class TurkorRelic : BossRelic
+    public class TurkorRelic : BossRelic
 	{
 		public override string RelicTextureName => "Consolaria/Content/Tiles/TurkorRelic";
 	}
